* use keyword arguments for files.File * fix: use Long insted of Integer type for file/dir sizes   we just removed 4GB limits, ahahaha * build number: start counting from 1 like normal human beings

+++ pisi/metadata.py
@@ -42,7 +42,7 @@ class Package(specfile.Package):
     t_Distribution = [ xmlfile.String, xmlfile.mandatory]
     t_DistributionRelease = [ xmlfile.String, xmlfile.mandatory]
     t_Architecture = [ xmlfile.String, xmlfile.mandatory]
-    t_InstalledSize = [ xmlfile.Integer, xmlfile.mandatory]
+    t_InstalledSize = [ xmlfile.Long, xmlfile.mandatory]
     t_PackageURI = [ xmlfile.String, xmlfile.optional]
 
     def decode_hook(self, node, errs, where):
